# web/swagger.yml
openapi: 3.0.1
info:
  title: Manta HTTP API
  version: 1.0.0

servers:
  - url: /api/v1
  # test only
  - url: http://localhost:8080/api/v1

tags:
  - name: Organizations
  - name: Otcls
    description: OpenTelemetry Collector

paths:
  /orgs:
    get:
      operationId: GetOrgs
      tags:
        - Organizations
      summary: List all organizations
      parameters:
        - $ref: "#/components/parameters/TraceSpan"
      responses:
        200:
          description: A list of organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organizations"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: CreateOrg
      tags:
        - Organizations
      summary: Create an Organization
      parameters:
        - $ref: "#/components/parameters/TraceSpan"
      requestBody:
        description: Organization to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
      responses:
        201:
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /orgs/{orgID}:
    get:
      operationId: GetOrgByID
      tags:
        - Organizations
      summary: Retrieve an organization
      parameters:
        - $ref: "#/components/parameters/TraceSpan"
        - in: path
          name: orgID
          schema:
            type: string
          required: true
          description: The ID of the organization to get
      responses:
        200:
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /otcls:
    get:
      operationId: GetOtcls
      tags:
        - Otcls
      summary: List all OpenTelemetry Collector
      parameters:
        - $ref: "#/components/parameters/TraceSpan"
      responses:
        200:
          description: A list of Otcl
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Otcl"
        default:
          $ref: "#/components/responses/UnexpectedError"

    post:
      operationId: CreateOtcl
      tags:
        - Otcls
      summary: Create an Otcl
      parameters:
        - $ref: "#/components/parameters/TraceSpan"
      requestBody:
        description: Otcl to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtclRequest"
      responses:
        201:
          description: Otcl created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Otcl"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /otcls/{otclID}:
    get:
      operationId: GetOtcl
      tags:
        - Otcls
      summary: Get Otcl by ID
      parameters:
        - $ref: "#"#/components/parameters/TraceSpan"
        - in: path
          name: otclID
          required: true
          description: The ID of the Otcl to get
      responses:
        200:
          description: Otcl details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Otcl"
        default:
          $ref: "#/components/responses/UnexpectedError"

components:
  parameters:
    TraceSpan:
      in: header
      name: Zap-Trace-Span
      description: OpenTracing span context
      example:
        trace_id: "1"
        span_id: "1"
        baggage:
          key: value
      required: false
      schema:
        type: string

  schemas:
    Error:
      properties:
        code:
          description: Code is the machine-readable error code.
          readOnly: true
          type: string
          # This set of enumerations must remain in sync with the constants defined in errors.go
          enum:
            - internal error
            - not found
            - conflict
            - invalid
            - unprocessable entity
            - empty value
            - unavailable
            - forbidden
            - too many requests
            - unauthorized
            - method not allowed
        message:
          readOnly: true
          description: Message is a human-readable message.
          type: string
      required: [ code, message ]

    Organization:
      type: object
      properties:
        id:
          readOnly: true
          type: string
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
        name:
          type: string
        desc:
          type: string
    Organizations:
      type: array
      items:
        $ref: "#/components/schemas/Organization"

    Otcl:
      type: object
      properties:
        id:
          readOnly: true
          type: string
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
        name:
          type: string
        desc:
          type: string
        orgID:
          type: string
          readOnly: true
        type:
          type: string
        content:
          type: string
          description: Config content for the OpenTelemetry Collector
    OtclRequest:
      type: object
      properties:
        name:
          type: string
        desc:
          type: string
        orgID:
          type: string
          readOnly: true
        type:
          type: string
        content:
          type: string
          description: Config content for the OpenTelemetry Collector

    Otcls:
      type: array
      items:
        $ref: "#/components/schemas/Otcl"

  responses:
    UnexpectedError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"